Create docker-compose.yml
To add a volume for the huginn service in the Docker Compose file you provided, you can add a new entry to the volumes field in the huginn service definition. For example:

  huginn:
    # other service configuration
    volumes:
      - "./data:/var/lib/MySQL"
      - "./huginn:/app/huginn"
This will create a new volume called huginn that is stored in a local directory called huginn (relative to the location of the Docker Compose file) and is mounted to the /app/huginn directory in the huginn container.

You may also want to consider adding a volume section to define the huginn volume, like so:

volumes:
  data:
  database:
  huginn:
This will create a named volume called huginn that can be reused across multiple services in your Docker Compose file.